+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 #### 🚀 Updates
 
+- Added an `--aliases` flag to `proto list` and `proto-list-remote`.
 - Updated `proto tool list` to include non-user aliases provided by the tool.
 - Updated `proto tool info` to include configuration and installed versions.